refactor: Moved book list item into its folder
